audio_hal: Apply the AudioRoute class in AudioHal
This patch is for applying the AudioRoute class in the
audio hal and solving following problems.
a) The Serializer parses the mixPort element in
audio_policy_configuration.xml as the AudioStreamRoute.
But it is not enough, because the MixPort can't represent
the audio backend route. So Serializer should parse mixPort
element as the MixPort, and parse the route element as the
AudioStreamRoute or AudioBackendRoute.

b) The route elements in audio_policy_configuration.xml
are parsed as Route struct, but it doesn't include enough
functions and data members. So now the route element will
be parsed as AudioRoute class, no longer Route struct.

c) RouteManagerConifg doesn't save the DevicePorts. It needs
to save the information of DevicePorts for both stream routes
and device routes

d) The isOut function can return the direction of a stream
route,0 for capture and 1 for playback. Its return value is
used as the index of array gOpenedRouteCriterion and
gRouteCriterionType to get the direction-specific criterion
and route types. But it It is not enough for supporting the
AudioBackendRoute, a backend route does not has a direction,
since it has two ports with different directions. So this patch
replace isOut with AudioRoute::getRouteType

Main changes:
a) Serializer (XML parser)
    1) Save the DevicePorts in RouteMangerConfig;
    2) Parse the mixPort XML element as the MixPorts
    rather than the AudioStreamRoute, and save them in
    RouteManagerConfig;
    3) Parse the route XML element as the AudioRoutes
    and save them in RouteMangerConfig;
b) IStreamRoute
    Change to a pure virtual interface. The getName and
    IsOut function is implemented by the derived class
    AudioStreamRoute. The AudioStreamRoute is derived
    from AudioRoute and IStreamRoute, and the AudioRoute
    also has a getName function. Since it isn't allowed to
    implement the same function in two base classes, so
    change the IStreamRoute to a pure virtual interface.
c) AudioStreamRoute
    1) Remove some data members and member functions, because
    the AudioRoute has implemented them;
    2) Remove the setConfig and setDevice fucntions, the config and
     AudioDevice have become assigned value in the constructor.
    3) Add the isOut function,  in order to implement the virtual
    interface of IStreamRoute::isOut. Although  audio route manage no
    longer needs it to get route types now, but audio stream manager
    still uses it in many files. So we keep this function.
    4) Add the getName function, in order to implement the virtual
    interface of IStreamRoute::getName.

Change-Id: Id5e49df7f32de71d688be58ea8227d79da700382
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>